perf(admin): load DevPass data client-side
The DevPass page awaited the subscriber list and the KPI strip server-side
before rendering anything, so the whole shell waited on the slowest
aggregate.

Split the KPI strip into its own GET /admin/devpass/kpis endpoint (its
ten aggregates now run concurrently), drop kpis from the list response,
and run the list rows and count query in parallel. The page itself
fetches nothing server-side: KPIs, chart, usage cards and the subscriber
table each load independently on the client with skeletons.
